LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/pseudo-elements-008-expected.txt

First-line
First-line

FAIL Originating element container for ::before assert_equals: expected "20px" but got "80px"
FAIL Originating element container for ::after assert_equals: expected "20px" but got "80px"
FAIL Originating element container for ::marker assert_equals: expected "20px" but got "80px"
FAIL Originating element container for ::first-line assert_equals: expected "20px" but got "80px"
FAIL Originating element container for ::first-letter assert_equals: expected "20px" but got "80px"
FAIL Originating element container  for ::highlight assert_equals: expected "20px" but got "80px"
FAIL Originating element container for outer ::first-line assert_equals: expected "30px" but got "80px"
FAIL Originating element container for outer ::first-letter assert_equals: expected "30px" but got "80px"
FAIL Originating element container for ::backdrop assert_equals: ::backdrop not rendered expected "30px" but got "80px"
